Ignore E402 module level import not at top of file
a790bcade9bf5544f81c1b0dbfe45161e34571e5 introduces conditional imports to support Wagtail 2.0
which causes the linter to complain about the following imports not being at the top of the file.

Ignoring the error on the following lines with `noqa` feels wrong as the error is not directly caused by it but by the imports above. Therefore ignoring all together for now. This can be removed one we drop support of Wagtail prior to 2.0.